+++ wcfsetup/install/files/lib/system/api/league/uri/BaseUri.php
@@ -33,6 +33,8 @@ use function implode;
 use function in_array;
 use function preg_match;
 use function rawurldecode;
+use function sort;
+use function str_contains;
 use function str_repeat;
 use function str_replace;
 use function strpos;
@@ -263,21 +265,32 @@ class BaseUri implements Stringable, JsonSerializable, UriAccess
      */
     public function isSameDocument(Stringable|string $uri): bool
     {
-        return self::normalizedUri($this->uri)->isSameDocument(self::normalizedUri($uri));
+        return self::normalizedUri($this->uri)->equals(self::normalizedUri($uri));
     }
 
     private static function normalizedUri(Stringable|string $uri): Uri
     {
-        $uri = ($uri instanceof Uri) ? $uri : Uri::new($uri);
-        $host = $uri->getHost();
-        if (null === $host || Ipv4Converter::fromEnvironment()->isIpv4($host) || IPv6Converter::isIpv6($host)) {
-            return $uri;
-        }
+        // Normalize the URI according to RFC3986
+        $uri = ($uri instanceof Uri ? $uri : Uri::new($uri))->normalize();
 
-        /** @var Uri $uri */
-        $uri = $uri->withHost(IdnaConverter::toUnicode((string) Ipv6Converter::compress($host))->domain());
+        return $uri
+            //Normalization as per WHATWG URL standard
+            //only meaningful for WHATWG Special URI scheme protocol
+            ->when(
+                condition: '' === $uri->getPath() && null !== $uri->getAuthority(),
+                onSuccess: fn (Uri $uri) => $uri->withPath('/'),
+            )
+            //Sorting as per WHATWG URLSearchParams class
+            //not included on any equivalence algorithm
+            ->when(
+                condition: null !== ($query = $uri->getQuery()) && str_contains($query, '&'),
+                onSuccess: function (Uri $uri) use ($query) {
+                    $pairs = explode('&', (string) $query);
+                    sort($pairs);
 
-        return $uri;
+                    return $uri->withQuery(implode('&', $pairs));
+                }
+            );
     }
 
     /**