+++ wcfsetup/install/files/lib/system/api/league/uri/Uri.php
@@ -21,7 +21,6 @@ use League\Uri\Contracts\FragmentDirective;
 use League\Uri\Contracts\UriComponentInterface;
 use League\Uri\Contracts\UriException;
 use League\Uri\Contracts\UriInterface;
-use League\Uri\Exceptions\ConversionFailed;
 use League\Uri\Exceptions\MissingFeature;
 use League\Uri\Exceptions\SyntaxError;
 use League\Uri\Idna\Converter as IdnaConverter;
@@ -43,22 +42,29 @@ use function array_filter;
 use function array_key_last;
 use function array_map;
 use function array_pop;
+use function array_shift;
 use function base64_decode;
 use function base64_encode;
 use function basename;
 use function count;
 use function dirname;
 use function explode;
+use function fclose;
 use function feof;
 use function file_get_contents;
 use function filter_var;
+use function fopen;
 use function fread;
+use function fwrite;
+use function gettype;
 use function implode;
 use function in_array;
-use function inet_pton;
 use function is_bool;
+use function is_object;
+use function is_resource;
 use function is_string;
 use function preg_match;
+use function preg_replace;
 use function preg_replace_callback;
 use function rawurldecode;
 use function rawurlencode;
@@ -79,7 +85,6 @@ use function trim;
 use const FILEINFO_MIME;
 use const FILEINFO_MIME_TYPE;
 use const FILTER_FLAG_IPV4;
-use const FILTER_FLAG_IPV6;
 use const FILTER_NULL_ON_FAILURE;
 use const FILTER_VALIDATE_BOOLEAN;
 use const FILTER_VALIDATE_EMAIL;
@@ -101,55 +106,11 @@ final class Uri implements Conditionable, UriInterface
     private const REGEXP_INVALID_CHARS = '/[\x00-\x1f\x7f]/';
 
     /**
-     * RFC3986 host identified by a registered name regular expression pattern.
-     *
-     * @link https://tools.ietf.org/html/rfc3986#section-3.2.2
-     *
-     * @var string
-     */
-    private const REGEXP_HOST_REGNAME = '/^(
-        (?<unreserved>[a-z\d_~\-\.])|
-        (?<sub_delims>[!$&\'()*+,;=])|
-        (?<encoded>%[A-F\d]{2})
-    )+$/x';
-
-    /**
-     * RFC3986 delimiters of the generic URI components regular expression pattern.
-     *
-     * @link https://tools.ietf.org/html/rfc3986#section-2.2
-     *
-     * @var string
-     */
-    private const REGEXP_HOST_GEN_DELIMS = '/[:\/?#\[\]@ ]/'; // Also includes space.
-
-    /**
-     * RFC3986 IPvFuture regular expression pattern.
-     *
-     * @link https://tools.ietf.org/html/rfc3986#section-3.2.2
-     *
-     * @var string
-     */
-    private const REGEXP_HOST_IP_FUTURE = '/^
-        v(?<version>[A-F\d])+\.
-        (?:
-            (?<unreserved>[a-z\d_~\-\.])|
-            (?<sub_delims>[!$&\'()*+,;=:])  # also include the : character
-        )+
-    $/ix';
-
-    /**
      * RFC3986 IPvFuture host and port component.
      *
      * @var string
      */
     private const REGEXP_HOST_PORT = ',^(?<host>(\[.*]|[^:])*)(:(?<port>[^/?#]*))?$,x';
-
-    /**
-     * Significant 10 bits of IP to detect Zone ID regular expression pattern.
-     *
-     * @var string
-     */
-    private const HOST_ADDRESS_BLOCK = "\xfe\x80";
 
     /**
      * Regular expression pattern to for file URI.
@@ -230,11 +191,17 @@ final class Uri implements Conditionable, UriInterface
         $this->pass = Encoder::encodePassword($pass);
         $this->host = $this->formatHost($host);
         $this->port = $this->formatPort($port);
+        $this->authority = UriString::buildAuthority([
+            'scheme' => $this->scheme,
+            'user' => $this->user,
+            'pass' => $this->pass,
+            'host' => $this->host,
+            'port' => $this->port,
+        ]);
         $this->path = $this->formatPath($path);
         $this->query = Encoder::encodeQueryOrFragment($query);
         $this->fragment = Encoder::encodeQueryOrFragment($fragment);
         $this->userInfo = null !== $this->pass ? $this->user.':'.$this->pass : $this->user;
-        $this->authority = UriString::buildAuthority($this->toComponents());
         $this->uriAsciiString = UriString::buildUri($this->scheme, $this->authority, $this->path, $this->query, $this->fragment);
         $this->assertValidRfc3986Uri();
         $this->assertValidState();
@@ -286,88 +253,7 @@ final class Uri implements Conditionable, UriInterface
      */
     private function formatHost(?string $host): ?string
     {
-        if (null === $host || '' === $host) {
-            return $host;
-        }
-
-        static $cache = [];
-        if (isset($cache[$host])) {
-            return $cache[$host];
-        }
-
-        $formattedHost = '[' === $host[0] ? $this->formatIp($host) : $this->formatRegisteredName($host);
-        $cache[$host] = $formattedHost;
-        if (self::MAXIMUM_CACHED_ITEMS < count($cache)) {
-            array_shift($cache);
-        }
-
-        return $formattedHost;
-    }
-
-    /**
-     * Validate and format a registered name.
-     *
-     * The host is converted to its ascii representation if needed
-     *
-     * @throws MissingFeature if the submitted host required missing or misconfigured IDN support
-     * @throws SyntaxError if the submitted host is not a valid registered name
-     * @throws ConversionFailed if the submitted IDN host cannot be converted to a valid ascii form
-     */
-    private function formatRegisteredName(string $host): string
-    {
-        $formattedHost = rawurldecode($host);
-        if ($formattedHost === $host) {
-            return match (1) {
-                preg_match(self::REGEXP_HOST_REGNAME, $formattedHost) => $formattedHost,
-                preg_match(self::REGEXP_HOST_GEN_DELIMS, $formattedHost) => throw new SyntaxError('The host `'.$host.'` is invalid : a registered name cannot contain URI delimiters or spaces.'),
-                default => IdnaConverter::toAsciiOrFail($host),
-            };
-        }
-
-        if (IdnaConverter::toAscii($formattedHost)->hasErrors()) {
-            throw new SyntaxError('The host `'.$host.'` is invalid : the registered name contains invalid characters.');
-        }
-
-        return (string) Encoder::normalizeHost($host);
-    }
-
-    /**
-     * Validate and Format the IPv6/IPvfuture host.
-     *
-     * @throws SyntaxError if the submitted host is not a valid IP host
-     */
-    private function formatIp(string $host): string
-    {
-        $ip = substr($host, 1, -1);
-        if (false !== filter_var($ip, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6)) {
-            return $host;
-        }
-
-        if (1 === preg_match(self::REGEXP_HOST_IP_FUTURE, $ip, $matches) && !in_array($matches['version'], ['4', '6'], true)) {
-            return $host;
-        }
-
-        $pos = strpos($ip, '%');
-        if (false === $pos) {
-            throw new SyntaxError('The host `'.$host.'` is invalid : the IP host is malformed.');
-        }
-
-        if (1 === preg_match(self::REGEXP_HOST_GEN_DELIMS, rawurldecode(substr($ip, $pos)))) {
-            throw new SyntaxError('The host `'.$host.'` is invalid : the IP host is malformed.');
-        }
-
-        $ip = substr($ip, 0, $pos);
-        if (false === filter_var($ip, FILTER_VALIDATE_IP, FILTER_FLAG_IPV6)) {
-            throw new SyntaxError('The host `'.$host.'` is invalid : the IP host is malformed.');
-        }
-
-        //Only the address block fe80::/10 can have a Zone ID attach to
-        //let's detect the link local significant 10 bits
-        if (str_starts_with((string)inet_pton($ip), self::HOST_ADDRESS_BLOCK)) {
-            return $host;
-        }
-
-        throw new SyntaxError('The host `'.$host.'` is invalid : the IP host is malformed.');
+        return HostRecord::from($host)->toAscii();
     }
 
     /**
@@ -620,8 +506,8 @@ final class Uri implements Conditionable, UriInterface
 
         return match ([]) {
             array_filter(explode(';', $parameters), $isInvalidParameter) => self::fromComponents([
-               'scheme' => 'data',
-               'path' => self::formatDataPath($mimetype.';'.$parameters.','.rawurlencode($data)),
+                'scheme' => 'data',
+                'path' => self::formatDataPath($mimetype.';'.$parameters.','.rawurlencode($data)),
             ]),
             default => throw new SyntaxError(sprintf('Invalid mediatype parameters, `%s`.', $parameters))
         };
@@ -797,11 +683,37 @@ final class Uri implements Conditionable, UriInterface
      */
     private function formatPath(string $path): string
     {
-        return match ($this->scheme) {
+        $path = match ($this->scheme) {
             'data' => Encoder::encodePath(self::formatDataPath($path)),
             'file' => self::formatFilePath(Encoder::encodePath($path)),
             default => Encoder::encodePath($path),
         };
+
+        if ('' === $path) {
+            return $path;
+        }
+
+        if (null !== $this->authority) {
+            // If there is an authority, the path must start with a `/`
+            return str_starts_with($path, '/') ? $path : '/'.$path;
+        }
+
+        // If there is no authority, the path cannot start with `//`
+        if (str_starts_with($path, '//')) {
+            return '/.'.$path;
+        }
+
+        $colonPos = strpos($path, ':');
+        if (false !== $colonPos && null === $this->scheme) {
+            // In the absence of a scheme and of an authority,
+            // the first path segment cannot contain a colon (":") character.'
+            $slashPos = strpos($path, '/');
+            (false !== $slashPos && $colonPos > $slashPos) || throw new SyntaxError(
+                'In absence of the scheme and authority components, the first path segment cannot contain a colon (":") character.'
+            );
+        }
+
+        return $path;
     }
 
     /**
@@ -903,7 +815,7 @@ final class Uri implements Conditionable, UriInterface
         }
 
         if (null === $this->authority && str_starts_with($this->path, '//')) {
-            throw new SyntaxError('If there is no authority the path `' . $this->path . '` cannot start with a `//`.');
+            throw new SyntaxError('If there is no authority the path `'.$this->path.'` cannot start with a `//`.');
         }
 
         $pos = strpos($this->path, ':');
@@ -917,7 +829,7 @@ final class Uri implements Conditionable, UriInterface
     }
 
     /**
-     * assert the URI scheme is valid
+     * assert the URI scheme is valid.
      *
      * @link https://w3c.github.io/FileAPI/#url
      * @link https://datatracker.ietf.org/doc/html/rfc2397
@@ -1398,6 +1310,36 @@ final class Uri implements Conditionable, UriInterface
         }
 
         return $host;
+    }
+
+    public function isIpv4Host(): bool
+    {
+        return HostRecord::isIpv4($this->host);
+    }
+
+    public function isIpv6Host(): bool
+    {
+        return HostRecord::isIpv6($this->host);
+    }
+
+    public function isIpvFutureHost(): bool
+    {
+        return HostRecord::isIpvFuture($this->host);
+    }
+
+    public function isIpHost(): bool
+    {
+        return HostRecord::isIp($this->host);
+    }
+
+    public function isRegisteredNameHost(): bool
+    {
+        return HostRecord::isRegisteredName($this->host);
+    }
+
+    public function isDomainHost(): bool
+    {
+        return HostRecord::isDomain($this->host);
     }
 
     public function getPort(): ?int