+++ wcfsetup/install/files/lib/system/api/pelago/emogrifier/src/HtmlProcessor/AbstractHtmlProcessor.php
@@ -4,7 +4,8 @@ declare(strict_types=1);
 
 namespace Pelago\Emogrifier\HtmlProcessor;
 
-use Pelago\Emogrifier\Utilities\Preg;
+use function Safe\preg_match;
+use function Safe\preg_replace;
 
 /**
  * Base class for HTML processor that e.g., can remove, add or modify nodes or attributes.
@@ -13,20 +14,13 @@ use Pelago\Emogrifier\Utilities\Preg;
  */
 abstract class AbstractHtmlProcessor
 {
-    /**
-     * @var non-empty-string
-     */
     protected const DEFAULT_DOCUMENT_TYPE = '<!DOCTYPE html>';
-
-    /**
-     * @var non-empty-string
-     */
     protected const CONTENT_TYPE_META_TAG = '<meta http-equiv="Content-Type" content="text/html; charset=utf-8">';
 
     /**
-     * @var non-empty-string Regular expression part to match tag names that PHP's DOMDocument implementation is not
-     *      aware are self-closing. These are mostly HTML5 elements, but for completeness `<command>` (obsolete) and
-     *      `<keygen>` (deprecated) are also included.
+     * Regular expression part to match tag names that PHP's DOMDocument implementation is not
+     * aware are self-closing. These are mostly HTML5 elements, but for completeness `<command>` (obsolete) and
+     * `<keygen>` (deprecated) are also included.
      *
      * @see https://bugs.php.net/bug.php?id=73175
      */
@@ -35,23 +29,17 @@ abstract class AbstractHtmlProcessor
     /**
      * Regular expression part to match tag names that may appear before the start of the `<body>` element.  A start tag
      * for any other element would implicitly start the `<body>` element due to tag omission rules.
-     *
-     * @var non-empty-string
      */
     protected const TAGNAME_ALLOWED_BEFORE_BODY_MATCHER
         = '(?:html|head|base|command|link|meta|noscript|script|style|template|title)';
 
     /**
      * regular expression pattern to match an HTML comment, including delimiters and modifiers
-     *
-     * @var non-empty-string
      */
     protected const HTML_COMMENT_PATTERN = '/<!--[^-]*+(?:-(?!->)[^-]*+)*+(?:-->|$)/';
 
     /**
      * regular expression pattern to match an HTML `<template>` element, including delimiters and modifiers
-     *
-     * @var non-empty-string
      */
     protected const HTML_TEMPLATE_ELEMENT_PATTERN
         = '%<template[\\s>][^<]*+(?:<(?!/template>)[^<]*+)*+(?:</template>|$)%i';
@@ -172,7 +160,7 @@ abstract class AbstractHtmlProcessor
         $htmlWithPossibleErroneousClosingTags = $this->getDomDocument()->saveHTML($this->getBodyElement());
         $bodyNodeHtml = $this->removeSelfClosingTagsClosingTags($htmlWithPossibleErroneousClosingTags);
 
-        return (new Preg())->replace('%</?+body(?:\\s[^>]*+)?+>%', '', $bodyNodeHtml);
+        return preg_replace('%</?+body(?:\\s[^>]*+)?+>%', '', $bodyNodeHtml);
     }
 
     /**
@@ -180,7 +168,7 @@ abstract class AbstractHtmlProcessor
      */
     private function removeSelfClosingTagsClosingTags(string $html): string
     {
-        return (new Preg())->replace('%</' . self::PHP_UNRECOGNIZED_VOID_TAGNAME_MATCHER . '>%', '', $html);
+        return preg_replace('%</' . self::PHP_UNRECOGNIZED_VOID_TAGNAME_MATCHER . '>%', '', $html);
     }
 
     /**
@@ -281,7 +269,7 @@ abstract class AbstractHtmlProcessor
     private function normalizeDocumentType(string $html): string
     {
         // Limit to replacing the first occurrence: as an optimization; and in case an example exists as unescaped text.
-        $result = (new Preg())->replace(
+        $result = preg_replace(
             '/<!DOCTYPE\\s++html(?=[\\s>])/i',
             '<!DOCTYPE html',
             $html,
@@ -309,17 +297,17 @@ abstract class AbstractHtmlProcessor
 
         // We are trying to insert the meta tag to the right spot in the DOM.
         // If we just prepended it to the HTML, we would lose attributes set to the HTML tag.
-        $hasHeadTag = (new Preg())->match('/<head[\\s>]/i', $html) !== 0;
+        $hasHeadTag = preg_match('/<head[\\s>]/i', $html) !== 0;
         $hasHtmlTag = \stripos($html, '<html') !== false;
 
         if ($hasHeadTag) {
-            $reworkedHtml = (new Preg())->replace(
+            $reworkedHtml = preg_replace(
                 '/<head(?=[\\s>])([^>]*+)>/i',
                 '<head$1>' . self::CONTENT_TYPE_META_TAG,
                 $html
             );
         } elseif ($hasHtmlTag) {
-            $reworkedHtml = (new Preg())->replace(
+            $reworkedHtml = preg_replace(
                 '/<html(.*?)>/is',
                 '<html$1><head>' . self::CONTENT_TYPE_META_TAG . '</head>',
                 $html
@@ -339,8 +327,46 @@ abstract class AbstractHtmlProcessor
      */
     private function hasContentTypeMetaTagInHead(string $html): bool
     {
-        (new Preg())->match(
-            '%^.*?(?=<meta(?=\\s)[^>]*\\shttp-equiv=(["\']?+)Content-Type\\g{-1}[\\s/>])%is',
+        preg_match(
+            '%
+                (?(DEFINE)
+                    # the target `http-equiv` attribute match
+                    (?<target_attribute>
+                        http-equiv=(["\']?+)Content-Type\\g{-1}
+                        # must be followed by one of these characters
+                        [\\s/>]
+                    )
+                    # the target `meta` element match without the opening `<`
+                    (?<target>
+                        meta(?=\\s)
+                        # one or other of these
+                        (?:
+                            # one or more characters other than `>` or space
+                            [^>\\s]++
+                            |
+                            # space not followed by the target `http-equiv` attribute
+                            \\s(?!(?&target_attribute))
+                        )
+                        # any number of times (including zero)
+                        *+
+                        \\s(?&target_attribute)
+                    )
+                )
+                # start of `subject`
+                ^
+                # one or other of these
+                (?:
+                    # one or more characters other than `<`
+                    [^<]++
+                    |
+                    # `<` not followed by `target`
+                    <(?!(?&target))
+                )
+                # any number of times (including zero)
+                *+
+                # followed by the target, not captured
+                (?=<(?&target))
+            %isx',
             $html,
             $matches
         );
@@ -369,10 +395,7 @@ abstract class AbstractHtmlProcessor
      */
     private function hasEndOfHeadElement(string $html): bool
     {
-        if (
-            (new Preg())->match('%<(?!' . self::TAGNAME_ALLOWED_BEFORE_BODY_MATCHER . '[\\s/>])\\w|</head>%i', $html)
-            !== 0
-        ) {
+        if (preg_match('%<(?!' . self::TAGNAME_ALLOWED_BEFORE_BODY_MATCHER . '[\\s/>])\\w|</head>%i', $html) !== 0) {
             // An exception to the implicit end of the `<head>` is any content within a `<template>` element, as well in
             // comments.  As an optimization, this is only checked for if a potential `<head>` end tag is found.
             $htmlWithoutCommentsOrTemplates = $this->removeHtmlTemplateElements($this->removeHtmlComments($html));
@@ -388,23 +411,19 @@ abstract class AbstractHtmlProcessor
     /**
      * Removes comments from the given HTML, including any which are unterminated, for which the remainder of the string
      * is removed.
-     *
-     * @throws \RuntimeException
      */
     private function removeHtmlComments(string $html): string
     {
-        return (new Preg())->throwExceptions(true)->replace(self::HTML_COMMENT_PATTERN, '', $html);
+        return preg_replace(self::HTML_COMMENT_PATTERN, '', $html);
     }
 
     /**
      * Removes `<template>` elements from the given HTML, including any without an end tag, for which the remainder of
      * the string is removed.
-     *
-     * @throws \RuntimeException
      */
     private function removeHtmlTemplateElements(string $html): string
     {
-        return (new Preg())->throwExceptions(true)->replace(self::HTML_TEMPLATE_ELEMENT_PATTERN, '', $html);
+        return preg_replace(self::HTML_TEMPLATE_ELEMENT_PATTERN, '', $html);
     }
 
     /**
@@ -413,7 +432,7 @@ abstract class AbstractHtmlProcessor
      */
     private function ensurePhpUnrecognizedSelfClosingTagsAreXml(string $html): string
     {
-        return (new Preg())->replace(
+        return preg_replace(
             '%<' . self::PHP_UNRECOGNIZED_VOID_TAGNAME_MATCHER . '\\b[^>]*+(?<!/)(?=>)%',
             '$0/',
             $html