+++ wcfsetup/install/files/lib/system/api/pelago/emogrifier/src/HtmlProcessor/CssToAttributeConverter.php
@@ -5,7 +5,10 @@ declare(strict_types=1);
 namespace Pelago\Emogrifier\HtmlProcessor;
 
 use Pelago\Emogrifier\Utilities\DeclarationBlockParser;
-use Pelago\Emogrifier\Utilities\Preg;
+
+use function Safe\preg_match;
+use function Safe\preg_replace;
+use function Safe\preg_split;
 
 /**
  * This HtmlProcessor can convert style HTML attributes to the corresponding other visual HTML attributes,
@@ -188,14 +191,12 @@ final class CssToAttributeConverter extends AbstractHtmlProcessor
      */
     private function mapWidthOrHeightProperty(\DOMElement $node, string $value, string $property): void
     {
-        $preg = new Preg();
-
         // only parse values in px and %, but not values like "auto"
-        if ($preg->match('/^(\\d+)(\\.(\\d+))?(px|%)$/', $value) === 0) {
+        if (preg_match('/^(\\d+)(\\.(\\d+))?(px|%)$/', $value) === 0) {
             return;
         }
 
-        $number = $preg->replace('/[^0-9.%]/', '', $value);
+        $number = preg_replace('/[^0-9.%]/', '', $value);
         $node->setAttribute($property, $number);
     }
 
@@ -245,8 +246,9 @@ final class CssToAttributeConverter extends AbstractHtmlProcessor
      */
     private function parseCssShorthandValue(string $value): array
     {
-        $values = (new Preg())->split('/\\s+/', $value);
+        $values = preg_split('/\\s+/', $value);
 
+        /** @var list<string> $values */
         $css = [];
         $css['top'] = $values[0];
         $css['right'] = (\count($values) > 1) ? $values[1] : $css['top'];