+++ wcfsetup/install/files/lib/system/api/pelago/emogrifier/src/HtmlProcessor/HtmlPruner.php
@@ -6,7 +6,9 @@ namespace Pelago\Emogrifier\HtmlProcessor;
 
 use Pelago\Emogrifier\CssInliner;
 use Pelago\Emogrifier\Utilities\ArrayIntersector;
-use Pelago\Emogrifier\Utilities\Preg;
+
+use function Safe\preg_match_all;
+use function Safe\preg_split;
 
 /**
  * This class can remove things from HTML.
@@ -18,8 +20,6 @@ final class HtmlPruner extends AbstractHtmlProcessor
      * supports XPath 1.0, lower-case() isn't available to us. We've thus far only set attributes to lowercase,
      * not attribute values. Consequently, we need to translate() the letters that would be in 'NONE' ("NOE")
      * to lowercase.
-     *
-     * @var non-empty-string
      */
     private const DISPLAY_NONE_MATCHER
         = '//*[@style and contains(translate(translate(@style," ",""),"NOE","noe"),"display:none")'
@@ -86,9 +86,8 @@ final class HtmlPruner extends AbstractHtmlProcessor
     {
         $classesToKeepIntersector = new ArrayIntersector($classesToKeep);
 
-        $preg = new Preg();
         foreach ($elements as $element) {
-            $elementClasses = $preg->split('/\\s++/', \trim($element->getAttribute('class')));
+            $elementClasses = preg_split('/\\s++/', \trim($element->getAttribute('class')));
             $elementClassesToKeep = $classesToKeepIntersector->intersectWith($elementClasses);
             if ($elementClassesToKeep !== []) {
                 $element->setAttribute('class', \implode(' ', $elementClassesToKeep));
@@ -123,11 +122,9 @@ final class HtmlPruner extends AbstractHtmlProcessor
      */
     public function removeRedundantClassesAfterCssInlined(CssInliner $cssInliner): self
     {
-        $preg = new Preg();
-
         $classesToKeepAsKeys = [];
         foreach ($cssInliner->getMatchingUninlinableSelectors() as $selector) {
-            $preg->matchAll('/\\.(-?+[_a-zA-Z][\\w\\-]*+)/', $selector, $matches);
+            preg_match_all('/\\.(-?+[_a-zA-Z][\\w\\-]*+)/', $selector, $matches);
             $classesToKeepAsKeys += \array_fill_keys($matches[1], true);
         }