+++ wcfsetup/install/files/lib/system/api/pelago/emogrifier/CHANGELOG.md
@@ -18,6 +18,24 @@ Please also have a look at our
 
 ### Fixed
 
+## 8.2.0: Add support for Symfony 8.0
+
+### Added
+
+- Add support for Symfony 8.0 (#1509)
+
+### Changed
+
+- Improve performance of checking for existing `Content-Type` (#1518, #1519)
+
+### Fixed
+
+- Support very large content in `meta` element in `head` (#1519)
+- Support very large HTML files without a `Content-Type` (#1518)
+- Allow numbers and underscores in CSS property names (#1500)
+- Use safe version of some PHP functions
+  (#1457, #1492, #1495, #1496, #1499, #1503, #1513, #1515, #1517)
+
 ## 8.1.0: Add support for PHP 8.5
 
 ### Added